+++ src/components/anim.js
@@ -1,0 +1,123 @@
+/* eslint-disable react/no-did-mount-set-state */
+
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { findDOMNode } from 'react-dom';
+import findKey from 'lodash/findKey';
+import { connect } from 'react-redux';
+import { VictoryAnimation } from 'victory-core';
+import { victoryEases } from '../utils/types';
+
+class Anim extends Component {
+  state = {
+    activeAnimation: -1
+  };
+
+  componentDidMount() {
+    const shouldDisableAnimation =
+      this.props.route.params.indexOf('export') !== -1 ||
+      this.props.route.params.indexOf('overview') !== -1;
+
+    if (shouldDisableAnimation) {
+      this.setState({ activeAnimation: this.props.toStyle.length - 1 });
+      return;
+    }
+
+    const order = this.props.order;
+    const node = findDOMNode(this.fragmentRef);
+    if (!node.dataset) {
+      node.dataset = {};
+    }
+    node.dataset.order = order;
+  }
+
+  componentWillReceiveProps(nextProps) {
+    const state = nextProps.fragment;
+    const slide = this.props.route.slide;
+    const fragment = findDOMNode(this.fragmentRef);
+    const slideHash = parseInt(this.context.slideHash);
+    const key = findKey(state.fragments[slide], {
+      id: `${slideHash}-${parseInt(fragment.dataset.fid)}`,
+    });
+
+    const shouldDisableAnimation =
+      nextProps.route.params.indexOf('export') !== -1 ||
+      nextProps.route.params.indexOf('overview') !== -1;
+
+    if (shouldDisableAnimation) {
+      this.setState({ activeAnimation: this.props.toStyle.length - 1 });
+      return;
+    }
+
+    if (
+      slide in state.fragments &&
+      state.fragments[slide].hasOwnProperty(key)
+    ) {
+      const animationStatus = state.fragments[slide][key].animations;
+      this.context.stepCounter.setFragments(state.fragments[slide], slide);
+      this.setState({
+        activeAnimation: animationStatus.every(a => a === true) ?
+          animationStatus.length - 1 :
+          animationStatus.indexOf(false) - 1
+      });
+    }
+  }
+
+  render() {
+    const {
+      children,
+      fromStyle,
+      toStyle,
+      transitionDuration,
+      easing,
+      style
+    } = this.props;
+    const child = React.Children.only(children);
+    const tweenData = this.state.activeAnimation === -1 ? fromStyle : toStyle[this.state.activeAnimation];
+    return (
+      <VictoryAnimation
+        data={tweenData}
+        duration={transitionDuration}
+        easing={easing}
+      >
+        {(tweenStyle) =>
+          React.cloneElement(child, {
+            className: `fragment ${child.props.className}`.trim(),
+            style: { ...child.props.style, ...style, ...tweenStyle },
+            ref: f => {
+              this.fragmentRef = f;
+            },
+            'data-animation-count': this.props.toStyle.length
+          })}
+      </VictoryAnimation>
+    );
+  }
+}
+
+Anim.defaultProps = {
+  order: 0
+};
+
+Anim.propTypes = {
+  children: PropTypes.node,
+  easing: PropTypes.oneOf(victoryEases).isRequired,
+  fragment: PropTypes.object,
+  fromStyle: PropTypes.object.isRequired,
+  order: PropTypes.number,
+  route: PropTypes.object,
+  style: PropTypes.object,
+  toStyle: PropTypes.arrayOf(PropTypes.object).isRequired,
+  transitionDuration: PropTypes.number.isRequired
+};
+
+Anim.contextTypes = {
+  export: PropTypes.bool,
+  overview: PropTypes.bool,
+  slide: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  slideHash: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  stepCounter: PropTypes.shape({
+    setFragments: PropTypes.func
+  })
+};
+
+export default connect(state => state)(Anim);
