+++ src/components/appear.js
@@ -2,99 +2,32 @@
 
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { findDOMNode } from 'react-dom';
-import findKey from 'lodash/findKey';
 import { connect } from 'react-redux';
-import { VictoryAnimation } from 'victory-core';
+import Anim from './anim';
+import { victoryEases } from '../utils/types';
 
 class Appear extends Component {
-  state = {
-    // active: false,
-    activeAnimation: -1
-  };
-
-  componentDidMount() {
-    const shouldDisableAnimation =
-      this.props.route.params.indexOf('export') !== -1 ||
-      this.props.route.params.indexOf('overview') !== -1;
-
-    if (shouldDisableAnimation) {
-      // this.setState({ active: true });
-      this.setState({ activeAnimation: this.props.endValue.length - 1 });
-      return;
-    }
-
-    const order = this.props.order || 0;
-    const node = findDOMNode(this.fragmentRef);
-    if (!node.dataset) {
-      node.dataset = {};
-    }
-    node.dataset.order = order;
-  }
-
-  componentWillReceiveProps(nextProps) {
-    const state = nextProps.fragment;
-    const slide = this.props.route.slide;
-    const fragment = findDOMNode(this.fragmentRef);
-    const slideHash = parseInt(this.context.slideHash);
-    const key = findKey(state.fragments[slide], {
-      id: `${slideHash}-${parseInt(fragment.dataset.fid)}`,
-    });
-
-    const shouldDisableAnimation =
-    nextProps.route.params.indexOf('export') !== -1 ||
-    nextProps.route.params.indexOf('overview') !== -1;
-
-    if (shouldDisableAnimation) {
-      // this.setState({ active: true });
-      this.setState({ activeAnimation: this.props.endValue.length - 1 });
-      return;
-    }
-
-    if (
-      slide in state.fragments &&
-      state.fragments[slide].hasOwnProperty(key)
-    ) {
-      // const active = state.fragments[slide][key].visible;
-      const animationStatus = state.fragments[slide][key].animations;
-      this.context.stepCounter.setFragments(state.fragments[slide], slide);
-      this.setState({
-        activeAnimation: animationStatus.every(a => a === true) ?
-          animationStatus.length - 1 :
-          animationStatus.indexOf(false) - 1
-      });
-    }
-  }
-
   render() {
     const {
-      children,
+      transitionDuration,
       startValue,
       endValue,
-      transitionDuration,
       easing,
       style
     } = this.props;
-    const child = React.Children.only(children);
-    const tweenData = this.state.activeAnimation === -1 ? startValue : endValue[this.state.activeAnimation];
     return (
-      <VictoryAnimation
-        data={tweenData}
-        duration={transitionDuration}
+      <Anim
+        transitionDuration={transitionDuration}
+        fromStyle={startValue}
+        toStyle={[endValue]}
         easing={easing}
+        style={style}
       >
-        {(tweenStyle) =>
-          React.cloneElement(child, {
-            className: `fragment ${child.props.className}`.trim(),
-            style: { ...child.props.style, ...style, ...tweenStyle },
-            ref: f => {
-              this.fragmentRef = f;
-            },
-            'data-animation-count': this.props.endValue.length
-          })}
-      </VictoryAnimation>
+        {this.props.children}
+      </Anim>
     );
   }
+
 }
 
 Appear.defaultProps = {
@@ -106,18 +39,7 @@ Appear.defaultProps = {
 
 Appear.propTypes = {
   children: PropTypes.node,
-  easing: PropTypes.oneOf([
-    'back', 'backIn', 'backOut', 'backInOut',
-    'bounce', 'bounceIn', 'bounceOut', 'bounceInOut',
-    'circle', 'circleIn', 'circleOut', 'circleInOut',
-    'linear', 'linearIn', 'linearOut', 'linearInOut',
-    'cubic', 'cubicIn', 'cubicOut', 'cubicInOut',
-    'elastic', 'elasticIn', 'elasticOut', 'elasticInOut',
-    'exp', 'expIn', 'expOut', 'expInOut',
-    'poly', 'polyIn', 'polyOut', 'polyInOut',
-    'quad', 'quadIn', 'quadOut', 'quadInOut',
-    'sin', 'sinIn', 'sinOut', 'sinInOut'
-  ]),
+  easing: PropTypes.oneOf(victoryEases),
   endValue: PropTypes.object,
   fragment: PropTypes.object,
   order: PropTypes.number,
