+++ src/utils/types.js
@@ -1,0 +1,12 @@
+export const victoryEases = [
+  'back', 'backIn', 'backOut', 'backInOut',
+  'bounce', 'bounceIn', 'bounceOut', 'bounceInOut',
+  'circle', 'circleIn', 'circleOut', 'circleInOut',
+  'linear', 'linearIn', 'linearOut', 'linearInOut',
+  'cubic', 'cubicIn', 'cubicOut', 'cubicInOut',
+  'elastic', 'elasticIn', 'elasticOut', 'elasticInOut',
+  'exp', 'expIn', 'expOut', 'expInOut',
+  'poly', 'polyIn', 'polyOut', 'polyInOut',
+  'quad', 'quadIn', 'quadOut', 'quadInOut',
+  'sin', 'sinIn', 'sinOut', 'sinInOut'
+];
